+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- [**BREAKING**] `pushFmt` now returns `SmlStrError!void`.
+
 ## [0.1.2] - 2023-07-22
 
 ### Added